# src/main/resources/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Swagger do Sistema de Votação em Pautas
  version: 1.0.0
  license:
    name: MIT
  contact:
    name: Pedro
    email: p.emerick@live.com
servers:
  - url: http://localhost:8085/v1
    description: Servidor de desenvolvimento local
paths:
  /agenda:
    post:
      summary: Criação de pauta para votação
      operationId: createAgenda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/AgendaDTO'
      responses:
        '201':
          description: Pauta criada com sucesso.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/AgendaDTO'
        '400':
          description: Erro na validação dos dados recebidos.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
  /agenda/{id}:
    get:
      summary: Busca de pautas por ID
      operationId: findAgendaById
      parameters:
        - name: id
          in: path
          description: ID da pauta
          required: true
          schema:
            type: integer
            format: int64
          style: simple
      responses:
        '200':
          description: Pauta encontrada.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/AgendaDTO'
        '404':
          description: Pauta não encontrada.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
  /agenda/{id}/session/start:
    post:
      summary: Iniciar sessão de uma pauta
      operationId: startSession
      parameters:
        - name: id
          in: path
          description: ID da pauta
          required: true
          schema:
            type: integer
            format: int64
          style: simple
        - name: duration
          in: query
          description: Minutos de duração da sessão, por padrão é definido um minuto
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Sessão de votação iniciada para a pauta com sucesso.
        '404':
          description: Pauta não encontrada.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '400':
          description: Erro na validação dos dados recebidos.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '409':
          description: Conflito com regra de negócio.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
  /agenda/{id}/vote:
    post:
      summary: Iniciar sessão de uma pauta
      operationId: vote
      parameters:
        - name: id
          in: path
          description: ID da pauta
          required: true
          schema:
            type: integer
            format: int64
          style: simple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/VoteEntryDTO'
      responses:
        '202':
          description: Votos recebidos para computação.
        '404':
          description: Pauta não encontrada.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '400':
          description: Erro na validação dos dados recebidos.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '409':
          description: Conflito com regra de negócio.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
  /agenda/{id}/result:
    get:
      summary: Visualização do resultado das votações
      operationId: getVoteResults
      parameters:
        - name: id
          in: path
          description: ID da pauta
          required: true
          schema:
            type: integer
            format: int64
          style: simple
      responses:
        '200':
          description: Resultados da pauta após sessão de votação.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/AgendaResultDTO'
        '404':
          description: Pauta não encontrada.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '400':
          description: Erro na validação dos dados recebidos.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'
        '409':
          description: Conflito com regra de negócio.
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/ErrorDTO'